+++ jsengine/jsengine-java/src/main/java/com/hyprmx/android/jsengine/JSEngineInvocationHandlerWrapper.java
@@ -1,0 +1,7 @@
+package com.hyprmx.android.jsengine;
+
+import java.lang.reflect.InvocationHandler;
+
+public interface JSEngineInvocationHandlerWrapper {
+    InvocationHandler wrapInvocationHandler(JavaScriptObject javaScriptObject, InvocationHandler handler);
+}