+++ jsengine/jsengine-java/src/main/java/com/hyprmx/android/jsengine/JSEngineJavaObject.java
@@ -1,0 +1,6 @@
+package com.hyprmx.android.jsengine;
+
+@SuppressWarnings("rawtypes")
+public interface JSEngineJavaObject {
+    Object getObject();
+}