+++ jsengine/jsengine-java/src/main/java/com/hyprmx/android/jsengine/JSEnginePromiseReceiver.java
@@ -1,0 +1,5 @@
+package com.hyprmx.android.jsengine;
+
+public interface JSEnginePromiseReceiver {
+    void receive(Object o);
+}