+++ jsengine/jsengine-java/src/main/java/com/hyprmx/android/jsengine/JSEngineProperty.java
@@ -1,0 +1,9 @@
+package com.hyprmx.android.jsengine;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface JSEngineProperty {
+    String name() default "";
+}