+++ jsengine/jsengine-java/src/main/java/com/hyprmx/android/jsengine/MemoizeFunc.java
@@ -1,0 +1,5 @@
+package com.hyprmx.android.jsengine;
+
+public interface MemoizeFunc<T> {
+  T process();
+}