+++ jsengine/jsengine-java/src/main/java/com/hyprmx/jsengine/jsengine/JSEngineJavaScriptObject.java
@@ -1,7 +1,0 @@
-package com.hyprmx.jsengine.jsengine;
-
-public interface JSEngineJavaScriptObject {
-    long getNativePointer();
-    long getNativeContext();
-    JavaScriptObject getJavaScriptObject();
-}