+++ jsengine/jsengine-java/src/main/java/com/hyprmx/jsengine/jsengine/JSEnginePromise.java
@@ -1,7 +1,0 @@
-package com.hyprmx.jsengine.jsengine;
-
-public interface JSEnginePromise {
-    JSEnginePromise then(JSEnginePromiseReceiver receiver);
-    @JSEngineMethodName(name = "catch")
-    JSEnginePromise caught(JSEnginePromiseReceiver receiver);
-}